requirements.txt
# ============================================
# AIOps Proactive Workflow - Dependencies
# Python 3.12+ required
# ============================================

# Core Agent Framework
strands-agents>=1.19.0
strands-agents-tools>=0.1.0

# AWS SDK
boto3>=1.35.0,<2.0.0
botocore>=1.35.0,<2.0.0

# AgentCore Runtime
bedrock-agentcore>=0.1.0

# Web API (Local Invocation Wrapper)
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
starlette>=0.36.0,<1.0.0

# Configuration Management
PyYAML>=6.0.1

# HTTP Client
requests>=2.31.0

# Data Validation
pydantic>=2.4.0,<3.0.0

# Environment Variables
python-dotenv>=1.0.0
